requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.1.0
Pillow==10.1.0
qrcode==7.4.2
WeasyPrint==60.1
gunicorn==21.2.0